% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_igd_create}
\alias{r_igd_create}
\title{Create an IGD database from a directory of bed files}
\usage{
r_igd_create(output_path, filelist, db_name)
}
\arguments{
\item{output_path}{String path where the IGD database will be saved}

\item{filelist}{String path to either a text file containing paths to bed files, or a directory containing bed files}

\item{db_name}{String name for the database (will be used in output filenames)}
}
\description{
Create an IGD database from a directory of bed files
}
